Integration/requirements.txt
defusedxml==0.4.1
Django==1.9.7
django-filter==0.13.0
djangorestframework==3.3.3
lxml==3.6.0
Markdown==2.6.6
python-dateutil==2.5.3
python-mimeparse==1.5.2
PyYAML==3.11
six==1.10.0
pygments